+++ app/calls/FeedbackFunctions.tsx
@@ -4,12 +4,16 @@ import { useUserData } from './UsersFunctions';
 type FeedbackEntity = {
     id: string;
     userId: string;
-    userName?: string;
-    userEmail?: string;
     feedbackTitle: string;
     feedbackDescription: string;
     createdAt: string;
+    user: user;
 };
+
+type user = {
+    name: string;
+    email: string;
+}
 
 const useFeedbackData = () => {
     const [feedback, setFeedback] = useState<FeedbackEntity[]>([]);
@@ -23,18 +27,7 @@ const useFeedbackData = () => {
                     throw new Error('Failed to fetch feedback data');
                 }
                 const feedbackData = await response.json();
-
-                // Update feedbackData with user email
-                const updatedFeedbackData = feedbackData.map((item: FeedbackEntity) => {
-                    const user = users.find(user => user.userId === item.userId);
-                    return {
-                        ...item,
-                        userEmail: user ? user.email : '',
-                        userName: user ? user.name : ''
-                    };
-                });
-
-                setFeedback(updatedFeedbackData);
+                setFeedback(feedbackData);
             } catch (error) {
                 console.error('Error fetching feedback data:', error);
             }
