+++ app/navs/Sidebar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import Link from 'next/link';
 
 interface SidebarProps {}
 
@@ -12,9 +12,9 @@ const Sidebar: React.FC<SidebarProps> = () => {
         <ul className="space-y-4 mt-2">
           <li>
             <Link
-              to="/dashboard"
+              href="/dashboard"
               style={{ color: '#344054' }}
-              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-orange-400 dark:hover:bg-orange-400 group"
+              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-[#FD9F10] dark:hover:bg-orange-400 group"
             >
               <svg className="w-6 h-6 transition-colors duration-100 ease-in-out group-hover:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                 <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="m4 12 8-8 8 8M6 10.5V19a1 1 0 0 0 1 1h3v-3a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v3h3a1 1 0 0 0 1-1v-8.5"/>
@@ -24,9 +24,9 @@ const Sidebar: React.FC<SidebarProps> = () => {
           </li>
           <li>
             <Link
-              to="/users"
+              href="/users"
               style={{ color: '#344054' }}
-              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-orange-400 dark:hover:bg-orange-400 group"
+              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-[#FD9F10] dark:hover:bg-orange-400 group"
             >
               <svg className="w-6 h-6 transition-colors duration-100 ease-in-out group-hover:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                 <path stroke="currentColor" strokeLinecap="round" strokeWidth="2" d="M16 19h4a1 1 0 0 0 1-1v-1a3 3 0 0 0-3-3h-2m-2.236-4a3 3 0 1 0 0-4M3 18v-1a3 3 0 0 1 3-3h4a3 3 0 0 1 3 3v1a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1Zm8-10a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z"/>
@@ -36,9 +36,9 @@ const Sidebar: React.FC<SidebarProps> = () => {
           </li>
           <li>
             <Link
-              to="/feedback"
+              href="/feedback"
               style={{ color: '#344054' }}
-              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-orange-400 dark:hover:bg-orange-400 group"
+              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-[#FD9F10] dark:hover:bg-orange-400 group"
             >
               <svg className="w-6 h-6 transition-colors duration-100 ease-in-out group-hover:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                 <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M12.01 6.001C6.5 1 1 8 5.782 13.001L12.011 20l6.23-7C23 8 17.5 1 12.01 6.002Z"/>
@@ -48,9 +48,9 @@ const Sidebar: React.FC<SidebarProps> = () => {
           </li>
           <li>
             <Link
-              to="/reports"
+              href="/reports"
               style={{ color: '#344054' }}
-              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-orange-400 dark:hover:bg-orange-400 group"
+              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-[#FD9F10] dark:hover:bg-orange-400 group"
             >
               <svg className="w-6 h-6 transition-colors duration-100 ease-in-out group-hover:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                 <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M10.779 17.779 4.36 19.918 6.5 13.5m4.279 4.279 8.364-8.643a3.027 3.027 0 0 0-2.14-5.165 3.03 3.03 0 0 0-2.14.886L6.5 13.5m4.279 4.279L6.499 13.5m2.14 2.14 6.213-6.504M12.75 7.04 17 11.28"/>
@@ -60,9 +60,9 @@ const Sidebar: React.FC<SidebarProps> = () => {
           </li>
           <li>
             <Link
-              to="/"
+              href="/"
               style={{ color: '#344054' }}
-              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-orange-400 dark:hover:bg-orange-400 group"
+              className="flex items-center p-3 text-gray-700 dark:text-white hover:bg-[#FD9F10] dark:hover:bg-orange-400 group"
             >
               <svg className="w-6 h-6 transition-colors duration-100 ease-in-out group-hover:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                 <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M20 12H8m12 0-4 4m4-4-4-4M9 4H7a3 3 0 0 0-3 3v10a3 3 0 0 0 3 3h2"/>
