+++ app/reports/page.tsx
@@ -1,47 +1,47 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useFeedbackData, useDeleteFeedback } from './calls/FeedbackFunctions';
-import Sidebar from './navs/Sidebar';
-import Navbar from './navs/Navbar';
+'use client'
 
-const Feedback = () => {
-    const allFeedbackItemsPerPage = 6;
-    const navigate = useNavigate();
-    const [allFeedbackCurrentPage, setFeedbackCurrentPage] = useState(1);
+import React, { useState } from 'react';
+import { useReportsData, useDeleteReport } from '../calls/ReportsFunctions';
+import Sidebar from '../navs/Sidebar';
+import Navbar from '../navs/Navbar';
+
+const Reports = () => {
+    const allReportsItemsPerPage = 6;
+    const [allReportsCurrentPage, setReportsCurrentPage] = useState(1);
     const [isVieModalVisible, setVieModalVisible] = useState(false);
     const [isDelModalVisible, setDelModalVisible] = useState(false);
-    const allFeedbackModalVisible = isVieModalVisible || isDelModalVisible;
-    const feedbackData = useFeedbackData();
-    const [selectedVieFeedbackId, setSelectedVieFeedbackId] = useState<string>("");
-    const [selectedDelFeedbackId, setSelectedDelFeedbackId] = useState<string>("");
-    const { deleteFeedback } = useDeleteFeedback(selectedDelFeedbackId);
+    const allReportsModalVisible = isVieModalVisible || isDelModalVisible;
+    const reportsData = useReportsData();
+    const [selectedDelReportId, setSelectedDelReportId] = useState<string>("");
+    const [selectedVieReportId, setSelectedVieReportId] = useState<string>("");
+    const { deleteReport } = useDeleteReport(selectedDelReportId);
 
     const toggleVieModal = (id: string) => {
-        setSelectedVieFeedbackId(id);
+        setSelectedVieReportId(id);
         setVieModalVisible(!isVieModalVisible);
     };
 
-    const selectedFeedback = feedbackData.find(feedback => feedback.id === selectedVieFeedbackId);
+    const selectedReport = reportsData.find(report => report.id === selectedVieReportId);
 
     const toggleDelModal = (id: string) => {
-        setSelectedDelFeedbackId(id);
+        setSelectedDelReportId(id);
         setDelModalVisible(!isDelModalVisible);
     };
 
-    const getFeedbackForCurrentPage = () => {
-        const startIndex = (allFeedbackCurrentPage - 1) * allFeedbackItemsPerPage;
-        const endIndex = startIndex + allFeedbackItemsPerPage;
-        return feedbackData.slice(startIndex, endIndex);
+    const getReportsForCurrentPage = () => {
+        const startIndex = (allReportsCurrentPage - 1) * allReportsItemsPerPage;
+        const endIndex = startIndex + allReportsItemsPerPage;
+        return reportsData.slice(startIndex, endIndex);
     };
 
-    const handleConfirmDeleteFeedback = async () => {
-        console.log(selectedDelFeedbackId);
+    const handleConfirmDeleteReport = async () => {
+        console.log(selectedDelReportId);
         try {
-            await deleteFeedback();
+            await deleteReport();
             setDelModalVisible(false);
-            navigate('/dashboard');
+            window.location.reload();
         } catch (error) {
-            console.error('Error deleting feedback:', error);
+            console.error('Error deleting report:', error);
         }
     };
 
@@ -50,13 +50,13 @@ const Feedback = () => {
                       height: '100vh', 
                       backgroundColor: '#F5F6FA' }}>
 
-            {allFeedbackModalVisible && (
-                <div className="f-overlay"></div>
+            {allReportsModalVisible && (
+                <div className="r-overlay"></div>
             )}
 
             <style>
                 {`
-                .f-overlay {
+                .r-overlay {
                     position: fixed;
                     top: 0;
                     left: 0;
@@ -66,14 +66,14 @@ const Feedback = () => {
                     z-index: 100;
                 }
 
-                .f-modal {
+                .r-modal {
                     position: fixed;
                     top: 31%;
                     left: 40%;
                     z-index: 101;
                 }
 
-                .f-modal-open {
+                .r-modal-open {
                     overflow: hidden;
                 }
                 `}
@@ -83,15 +83,15 @@ const Feedback = () => {
             <Navbar />
 
             <div
-                id="f-modal"
-                className={`f-modal ${isVieModalVisible ? '' : 'hidden'}`}
+                id="r-modal"
+                className={`r-modal ${isVieModalVisible ? '' : 'hidden'}`}
             >
                 <div className="fixed top-0 left-0 w-full h-full flex justify-center items-center z-50">
                     <div className="w-[600px] h-[350px] bg-white rounded-xl">
                         {/* Header section */}
-                        <div className="bg-[#FD9F10] rounded-t-lg text-white p-6 mb-4 flex justify-between items-center">
+                        <div className="bg-[#FD9F10] rounded-t-lg text-white p-6 flex justify-between items-center">
                             <h2 className="text-2xl font-poppins font-semibold ml-2">
-                                Feedback Details
+                                Report Details
                             </h2>
                             <button
                                 onClick={() => {
@@ -108,38 +108,45 @@ const Feedback = () => {
                                      fill="none" 
                                      viewBox="0 0 24 24">
                                     <path stroke="currentColor" 
-                                          stroke-linecap="round" 
-                                          stroke-linejoin="round" 
-                                          stroke-width="2" 
+                                          strokeLinecap="round" 
+                                          strokeLinejoin="round" 
+                                          strokeWidth="2" 
                                           d="m15 9-6 6m0-6 6 6m6-3a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z"/>
                                 </svg>
                             </button>
                         </div>
-                        {/* Feedback Details */}
-                        <div className='p-8 pt-2 font-poppins text-[#344054] mt-2'>
-                            <div className="flex flex-row items-center justify-between">
+                        {/* Report Details */}
+                        <div className='p-8 pt-2 font-poppins text-[#344054] overflow-y-auto' style={{ maxHeight: '250px' }}>
+                            <div className="flex flex-row items-center justify-between mt-5">
                                 <div className="flex flex-col">
                                     <h1 className="text-xl font-bold">
                                         Title
                                     </h1>
                                     <p className="mt-2 text-sm">
-                                        {selectedFeedback?.feedbackTitle}
+                                        {selectedReport?.reportTitle}
                                     </p>
                                 </div>
                                 <div className="flex flex-col">
                                     <h1 className="text-xl font-bold">
-                                        Date Sent
+                                        Submission
                                     </h1>
                                     <p className="mt-2 text-sm">
-                                        {selectedFeedback ? new Date(selectedFeedback.createdAt).toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' }) : ''}
+                                        {selectedReport ? 
+                                                new Date(selectedReport.createdAt).toLocaleString('en-US', {
+                                                month: 'long',
+                                                day: 'numeric',
+                                                year: 'numeric',
+                                                hour: 'numeric',
+                                                minute: 'numeric'
+                                                }) : ''}
                                     </p>
                                 </div>
                             </div>
                             <h1 className="text-xl font-bold mt-6">
-                                Feedback
+                                Report
                             </h1>
                             <p className="mt-2 text-sm">
-                                {selectedFeedback?.feedbackDescription}
+                                {selectedReport?.reportDescription}
                             </p>
                         </div>
                     </div>
@@ -147,18 +154,18 @@ const Feedback = () => {
             </div>
 
             <div
-                id="f-modal"
-                className={`f-modal ${isDelModalVisible ? '' : 'hidden'}`}
+                id="r-modal"
+                className={`r-modal ${isDelModalVisible ? '' : 'hidden'}`}
             >
                 <div className="w-full max-w-md max-h-full"
-                    style={{width: '300px'}}>
+                     style={{width: '300px'}}>
                     <div className="relative bg-white border border-white rounded-xl">
                         <div className="flex flex-col items-center mt-4 mb-4 p-4">
                             <div className="mt-2 font-bold font-poppins text-xl text-[#CF453B]">
-                                Delete Feedback
+                                Delete Report
                             </div>
                             <div className="font-semibold font-poppins text-md text-center text-[#344054] p-4">
-                                Are you sure you want to delete the feedback?
+                                Are you sure you want to delete the report?
                             </div>
                             <div className="flex flex-row mt-2 mb-2">
                                 <button type="button" 
@@ -168,7 +175,7 @@ const Feedback = () => {
                                     Cancel
                                 </button>
                                 <button type="button" 
-                                        onClick={() => handleConfirmDeleteFeedback()} 
+                                        onClick={() => handleConfirmDeleteReport()} 
                                         style={{ width: '120px' }} 
                                         className="px-8 py-2 rounded text-[#CF453B] bg-[#F9DAD8] hover:bg-[#E2C7C5] hover:text-[#B83F36] font-semibold font-poppins cursor-pointer text-center text-sm">
                                     Delete
@@ -185,12 +192,12 @@ const Feedback = () => {
                              fontSize: '30px', 
                              marginTop: '100px', 
                              marginLeft: '300px' }}>
-                    Feedback
+                    Reports
                 </h2>
                 <div style={{ marginLeft: '300px', 
                               marginTop: '20px' }} 
                      className="p-0 pb-0 pt-4">
-                    {/* Feedback Table */}
+                    {/* Reports Table */}
                     <table style={{ width: '1180px' }}>
                         <colgroup>
                             <col style={{ width: '300px' }} />
@@ -202,7 +209,7 @@ const Feedback = () => {
                                         backgroundColor: '#FFFFFF', 
                                         fontSize: '16px' }} 
                                className="font-poppins font-bold shadow-md">
-                            <tr style={{  height: '60px' }}>
+                            <tr style={{ height: '60px' }}>
                                 <th scope="col" 
                                     className="px-5 py-3 text-left">
                                     Username
@@ -213,7 +220,7 @@ const Feedback = () => {
                                 </th>
                                 <th scope="col" 
                                     className="px-3 py-3 text-center">
-                                    Feedback
+                                    Report
                                 </th>
                                 <th scope="col" 
                                     className="px-3 py-3 text-center">
@@ -222,10 +229,10 @@ const Feedback = () => {
                             </tr>
                         </thead>
                         <tbody>
-                            {getFeedbackForCurrentPage().map((feedback, index) => (
-                                <React.Fragment key={feedback.id}>
+                            {getReportsForCurrentPage().map((reports, index) => (
+                                <React.Fragment key={reports.id}>
                                     {/* Empty row for spacing */}
-                                    {index !== getFeedbackForCurrentPage().length && (
+                                    {index !== getReportsForCurrentPage().length && (
                                         <tr style={{ height: '10px' }}></tr>
                                     )}
                                     <tr style={{ color: '#344054', 
@@ -234,16 +241,16 @@ const Feedback = () => {
                                         className="font-poppins shadow-md">
                                         <td scope="row" 
                                             className="px-5 py-3 text-left">
-                                            {feedback.userName}
+                                            {reports.user.name}
                                         </td>
                                         <td scope="row" 
                                             className="px-5 py-3 text-left">
-                                            {feedback.userEmail}
+                                            {reports.user.email}
                                         </td>
                                         <td scope="row" 
                                             className="px-3 py-3 text-center">
                                             <button
-                                                onClick={() => toggleVieModal(feedback.id)}
+                                                onClick={() => toggleVieModal(reports.id)}
                                                 style={{ background: 'none', 
                                                          border: 'none', 
                                                          padding: 0, 
@@ -255,12 +262,13 @@ const Feedback = () => {
                                                 onMouseEnter={(e) => (e.currentTarget.style.textDecoration = 'underline')}
                                                 onMouseLeave={(e) => (e.currentTarget.style.textDecoration = 'none')}
                                             >
-                                                View Feedback
+                                                View Report
                                             </button>
                                         </td>
-                                        <td scope="row" className="py-2 flex items-center justify-center">
+                                        <td scope="row" 
+                                            className="py-2 flex items-center justify-center">
                                             <button
-                                                onClick={() => toggleDelModal(feedback.id)}
+                                                onClick={() => toggleDelModal(reports.id)}
                                                 style={{ width: '80px' }}
                                                 className="focus:outline-none text-xs text-[#CF453B] bg-[#F9DAD8] hover:bg-[#E2C7C5] hover:text-[#B83F36] font-medium font-poppins rounded-lg px-5 py-2 me-2 mb-1 mt-1 ml-2"
                                             >
@@ -272,14 +280,14 @@ const Feedback = () => {
                             ))}
                         </tbody>
                     </table>
-                    {/* Pagination for Feedback */}
+                    {/* Pagination for Reports */}
                     <div className="mt-2 p-2 rounded-lg">
-                        {Array.from({ length: Math.ceil(feedbackData.length / allFeedbackItemsPerPage) }, (_, index) => (
+                        {Array.from({ length: Math.ceil(reportsData.length / allReportsItemsPerPage) }, (_, index) => (
                             <button key={index + 1}
                                     className={`mx-1 px-3 py-2 text-xs font-bold rounded-full ${
-                                        allFeedbackCurrentPage === index + 1 ? 'bg-[#344054] text-white' : 'bg-[#E6E6E6] text-[#4C4C4C] hover:bg-[#344054] hover:text-white'
+                                        allReportsCurrentPage === index + 1 ? 'bg-[#344054] text-white' : 'bg-[#E6E6E6] text-[#4C4C4C] hover:bg-[#344054] hover:text-white'
                                     }`}
-                                    onClick={() => setFeedbackCurrentPage(index + 1)}>
+                                    onClick={() => setReportsCurrentPage(index + 1)}>
                                 {index + 1}
                             </button>
                         ))}
@@ -290,4 +298,4 @@ const Feedback = () => {
     );
 };
 
-export default Feedback;
+export default Reports;
